Cleaning up a bit more: moving config stuff to config folder
